+++ auth/grpctransport/directpath.go
@@ -16,22 +16,15 @@ package grpctransport
 
 import (
 	"context"
-	"log"
 	"net"
 	"os"
-	"runtime"
 	"strconv"
 	"strings"
-	"sync"
 
 	"cloud.google.com/go/auth"
+	"cloud.google.com/go/internal/compute"
 	"google.golang.org/grpc"
 	grpcgoogle "google.golang.org/grpc/credentials/google"
-)
-
-var (
-	vmOnGCEOnce sync.Once
-	vmOnGCE     bool
 )
 
 func isDirectPathEnabled(endpoint string, opts *Options) bool {
@@ -98,7 +91,7 @@ func isDirectPathXdsUsed(o *Options) bool {
 // configuration allows the use of direct path. If it does not the provided
 // grpcOpts and endpoint are returned.
 func configureDirectPath(grpcOpts []grpc.DialOption, opts *Options, endpoint string, creds *auth.Credentials) ([]grpc.DialOption, string) {
-	if isDirectPathEnabled(endpoint, opts) && OnComputeEngine() && isTokenProviderDirectPathCompatible(creds, opts) {
+	if isDirectPathEnabled(endpoint, opts) && compute.OnComputeEngine() && isTokenProviderDirectPathCompatible(creds, opts) {
 		// Overwrite all of the previously specific DialOptions, DirectPath uses its own set of credentials and certificates.
 		grpcOpts = []grpc.DialOption{
 			grpc.WithCredentialsBundle(grpcgoogle.NewDefaultCredentialsWithOptions(grpcgoogle.DefaultCredentialsOptions{PerRPCCreds: &grpcCredentialsProvider{creds: creds}}))}
@@ -127,38 +120,4 @@ func configureDirectPath(grpcOpts []grpc.DialOption, opts *Options, endpoint str
 		// TODO: add support for system parameters (quota project, request reason) via chained interceptor.
 	}
 	return grpcOpts, endpoint
-}
-
-// OnComputeEngine returns whether the client is running on GCE.
-//
-// It is a copy of the gRPC internal OnGCE() func at:
-// https://github.com/grpc/grpc-go/blob/master/internal/googlecloud/googlecloud.go
-func OnComputeEngine() bool {
-	vmOnGCEOnce.Do(func() {
-		mf, err := manufacturer()
-		if err != nil {
-			log.Printf("Failed to read manufacturer, setting onGCE=false: %v", err)
-			return
-		}
-		vmOnGCE = isRunningOnGCE(mf, runtime.GOOS)
-	})
-	return vmOnGCE
-}
-
-// isRunningOnGCE checks whether the local system, without doing a network request, is
-// running on GCP.
-func isRunningOnGCE(manufacturer []byte, goos string) bool {
-	name := string(manufacturer)
-	switch goos {
-	case "linux":
-		name = strings.TrimSpace(name)
-		return name == "Google" || name == "Google Compute Engine"
-	case "windows":
-		name = strings.Replace(name, " ", "", -1)
-		name = strings.Replace(name, "\n", "", -1)
-		name = strings.Replace(name, "\r", "", -1)
-		return name == "Google"
-	default:
-		return false
-	}
 }
