+++ internal/compute/manufacturer.go
@@ -1,3 +1,6 @@
+//go:build !(linux || windows)
+// +build !linux,!windows
+
 // Copyright 2024 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,12 +15,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package grpctransport
-
-import "os"
-
-const linuxProductNameFile = "/sys/class/dmi/id/product_name"
+package compute
 
 func manufacturer() ([]byte, error) {
-	return os.ReadFile(linuxProductNameFile)
+	return nil, nil
 }
