+++ internal/compute/manufacturer_linux.go
@@ -1,6 +1,3 @@
-//go:build !(linux || windows)
-// +build !linux,!windows
-
 // Copyright 2024 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,8 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package grpctransport
+package compute
+
+import "os"
+
+const linuxProductNameFile = "/sys/class/dmi/id/product_name"
 
 func manufacturer() ([]byte, error) {
-	return nil, nil
+	return os.ReadFile(linuxProductNameFile)
 }
