+++ internal/compute/manufacturer_windows.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package grpctransport
+package compute
 
 import (
 	"errors"
